Fix Bootcamp client interactions and scene recovery
Refresh the initial weapon tray and repair starting-pistol metadata. Preserve 64-bit consumable item identities and apply medpacks with atomic single consumption. Add the mission-inventory bomb with pickup, planting, retry and legacy recovery, and publish enabled scene objects as targetable. Defer valid static actor spawning during scene recovery while retaining genuine failure diagnostics.

+++ src/Rasa.Game/Structures/ActionData.cs
@@ -16,7 +16,7 @@ namespace Rasa.Structures
         /// <summary>Where a ground-targeted ability was aimed, when it had no target entity.</summary>
         public Vector3? TargetLocation { get; set; }
 
-        public int ItemId { get; set; }
+        public ulong ItemId { get; set; }
         public long WaitTime { get; set; }
         public long PassedTime { get; set; }
